针对大量文本处理进行优化
解决转化大量文本出现的乱码以及清洗时出现的问题
