--- test/filelist.txt
@@ -1,38 +1,0 @@
-%0 Journal Article
-%A 刘自强王效岳白如江
-%D 2016
-%T 多维度视角下学科主题演化可视化分析方法研究——以我国图书情报领域大数据研究为例
-%J 中国图书馆学报
-%N 06
-%U http://www.cnki.net/kcms/download.aspx?filename=EcQ9yU2EGUXpXWRJVa4plVlJ1dkdWVQpENjZlQmpVTLVzdSxEUyNGVvl0UFp2SnlXbqFkV2l3RFlGS=0zbZ5GMq1GcOpHbNZEdqBnN2lnQwE2ZadlcIdXc4QWR0ZTUkhEV2ZUaCdlaS1keywEU2YDUUV0S3d&tablename=CJFDLAST2016&dflag=pdfdown 全文链接
-
-%H Wangxiaoyue Liu Zi Qiang
-%W CNKI
-
-
-
-%0 Journal Article
-%A 王效岳白如江王晓笛祝娜
-%D 2013
-%T 海量网络学术文献自动分类系统
-%J 图书情报工作
-%N 16
-%U http://www.cnki.net/kcms/download.aspx?filename=wckYlNWdWBTVHxGdw0GOrI2ZxUUQ3ZmRIRjenJnTot2arFHSzpWQW1UZThXWT1WY4I1ZYdGVK5mdZtGdJpkd3JnVKV2UrR1VqdXRQhXaVRHd6VFbiZjWUFjYwRWVNJXe6ZnZVVXbEl0duFVMpZXNKd3KllWajJHc&tablename=CJFDHIS2&dflag=pdfdown 全文链接
-
-%H Bairujiang Wang Xiao Yue
-%W CNKI
-
-
-
-%0 Journal Article
-%A 胡珊
-%D 2012
-%T 基于CSSCI计量分析我国学科信息门户的发展及语义网在学科信息门户中的应用研究
-%J 现代情报
-%N 07
-%U http://www.cnki.net/kcms/download.aspx?filename=DVFODVlV3RFazMWQrI2TpV3RNhGaohkMplTTqZ0QTFTZwQVT5hHbzQlYxFEazBnRLhUeJhlRpFGRW9UQ2cEcYt0axlHeVxET1hmRvo2Z0pVdid0Y2dWYwgTY2E0dhJ3Q59CbjhGaoZlYrB1V3k0QF5UNpRERqhjM&tablename=CJFD2012&dflag=pdfdown 全文链接
-
-%H Shan Hu
-%W CNKI
-
-
